refactor #60: 상품 삭제 api 훅으로 분리
