+++ fe/src/api/product/queries.ts
@@ -1,5 +1,11 @@
-import { useInfiniteQuery, useQuery } from "@tanstack/react-query";
-import { getAddresses, getCategories, getProductDetail } from ".";
+import { useInfiniteQuery, useMutation, useQuery } from "@tanstack/react-query";
+import {
+  deleteProduct,
+  getAddresses,
+  getCategories,
+  getProductDetail,
+} from ".";
+import { useToast } from "@hooks/useToast";
 
 export const useAddressesInfiniteQuery = (option?: { size: number }) => {
   return useInfiniteQuery({
@@ -25,4 +31,31 @@ export const useProductDetailQuery = (id: number, enabled: boolean) => {
     enabled,
     staleTime: Infinity,
   });
+};
+
+export const useDeleteProductQuery = (productId: number) => {
+  const { toast } = useToast();
+
+  const deleteProductMutation = useMutation(deleteProduct);
+
+  const onDeleteProduct = () => {
+    deleteProductMutation.mutate(productId, {
+      onSuccess: () => {
+        toast({
+          type: "success",
+          title: "상품 삭제 성공",
+          message: "상품 삭제에 성공했습니다.",
+        });
+      },
+      onError: () => {
+        toast({
+          type: "error",
+          title: "상품 삭제 실패",
+          message: "상품 삭제에 실패했습니다. 잠시 후 다시 시도해주세요.",
+        });
+      },
+    });
+  };
+
+  return { onDeleteProduct };
 };
